pyproject.toml
[project]
name = "josafa"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "nayetdet", email = "jpmoreiracampos21@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.0",
    "discord-py[voice]>=2.6.3",
    "motor>=3.7.1",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.9.5",
]

[tool.uv.workspace]
members = []
